move shift_months test to test_arithmetic
